# cpu/stm32_common: EEPROM speedup