--- a/cpu/stm32_common/periph/flash_common.c
+++ b/cpu/stm32_common/periph/flash_common.c
@@ -56,3 +56,23 @@ void _lock(void)
         CNTRL_REG |= CNTRL_REG_LOCK;
     }
 }
+
+void _wait_for_pending_operations(void)
+{
+    if ((FLASH->SR & FLASH_SR_BSY) == FLASH_SR_BSY) {
+        DEBUG("[flash-common] waiting for any pending operation to finish\n");
+        while (FLASH->SR & FLASH_SR_BSY) {}
+    }
+    else {
+        if ((FLASH->SR & (uint32_t)FLASH_SR_WRPERR) != (uint32_t)0x00) {
+            DEBUG("[flash-common] resettng previously occured flash error\n");
+            FLASH->SR |= (uint32_t)FLASH_SR_WRPERR;
+            while (FLASH->SR & FLASH_SR_BSY) {}
+        }
+    }
+
+    /* Clear 'end of operation' bit in status register */
+    if (FLASH->SR & FLASH_SR_EOP) {
+        FLASH->SR &= ~(FLASH_SR_EOP);
+    }
+}