--- a/tests/periph_eeprom/Makefile
+++ b/tests/periph_eeprom/Makefile
@@ -3,6 +3,7 @@ include ../Makefile.tests_common
 
 FEATURES_REQUIRED += periph_eeprom
 
+USEMODULE += xtimer
 USEMODULE += shell
 USEMODULE += shell_commands  # provides reboot command