### Contribution description

Changing EEPROM access from byte to 32-bit word improves access speed by up to 8 times. Unaligned and less than 4 bytes access still possible using same functions.

P.S. _wait_for_pending_operations moved from flashpage to flash_common and used in eeprom.c to prevent flash controller busy errors.

### Testing procedure

'test' command was added to tests/periph_eeprom

Before (STM32L151CCU6 @ 32 MHz):
Writing 256 bytes to clean EEPROM: 1709029 usec
Writing 256 bytes to dirty EEPROM: 1728992 usec
Reading 256 bytes from EEPROM: 408 usec

After:
Writing 256 bytes to clean EEPROM: 216165 usec
Writing 256 bytes to dirty EEPROM: 432014 usec
Reading 256 bytes from EEPROM: 242 usec