@olegart 2018-09-25
Ok, so

* comments noted. I left __IO modifier in place - usually it translates to 'volatile', and it's a bit safer to have volatile access to hardware storage. At least, no funny compilers' optimizations to expect.

* read/write functions further improved - there's no less-than-word access anymore. Accessing single bytes are really expensive and translates to multiple read-modify-clear-write operations, wearing eeprom out. All accesses with eeprom_read and eeprom_write are aligned 1-word operations on hardware level now.

* code tested on our production firmware (stores hundreds of bytes of data in EEPROM in chunks from 1 byte to few hundreds bytes in size), it works ok. Someone may want to check aligned / unaligned access with different data sizes and start addresses, but as I see it now, everything is ok.

* there will be no problems with flash_common - there's a single function implementation had been moved from flashpage to flash_common. FLASH->SR must be checked for BUSY and ERROR flags before any flash *or EEPROM* read/write operation, so _wait_for_pending_operations() should never be limited to flashpage.c in the first place.

* actually, eeprom_clear implementation is already in place - just call eeprom_write with NULL instead of data pointer.