@aabadie 2018-09-25
Thanks for addressing my comments quickly @olegart. I may reconsider my request about contributing the clear functions in a separate PR.

As you can see in the PR status, some static check are still failing on Travis. They are about commit messages too long and trailing whitespaces.

Please fix them in fixup commits. We want the git history to remain as clean as possible, so as I see this PR now, you'll have to rework the git history. Let me know if this causes you problems, I can help.