@aabadie 2018-09-26
Thanks @olegart, it looks super good now. I'll test the PR tomorrow on a couple of boards.

By the way, you removed the clear functions, do you plan to open another PR to add them ? It can be based on this one if you want.