@olegart 2018-09-27
@aabadie, yep, clear functions needs to be added to atmega_common as well, so separate PR for it.