@aabadie 2018-09-27
> clear functions needs to be added to atmega_common as well, so separate PR for it

I think there's no need to modify the CPU specific code for clear functions, see #10056. I also fixed there the bounds issues (last byte in EEPROM couldn't be read/written)