@aabadie 2018-09-27
@olegart, after working on #10056, I had a second look at your changes and I see an important conceptual problem in your PR (sorry I missed it initially): you are reimplementing the eeprom_read/write in the STM32 common periph code. I'm wondering why it builds since this function is already defined in `drivers/periph_common/eeprom.c`.
You are also introducing internal `_eeprom_read_word` and `_eeprom_write_word` functions that are used to speed up global `eeprom_read/write` functions instead of the slow `eeprom_read/write_byte` one.
`word` is a name (weirdly) used by ST for 4 bytes values, on other architectures this is in general called dword (double word), and word is used for values on 2 bytes.

The actual API provides 4 functions:  eeprom_read, eeprom_write, eeprom_read_byte and eeprom_write_byte.

So I see 3 alternatives:
1. Keep the public API unchanged, implement a fast eeprom_read and eeprom_write and skip the ones in periph_common using defines only for STM32. AVR will provide their own slow (and EEPROM unsafe) implementations. This is basically what this PR does but it misses the defines to skip the implementation in periph_common. Even if it builds, it would make things cleaner.
The drawback of this method is that it duplicates more code (more functions defined in different places).
2. Add `eeprom_read_dword`, `eeprom_write_dword` function to the public API in drivers/periph_eeprom, and modify the public implementations of eeprom_read, eeprom_write like it's done in this PR. That would require the 2 new functions for AVR (but it can be a simple loop with a call to eeprom_read_byte, eeprom_write_byte. The drawback of this solution is that it will increase the code size of application built with periph_eeprom and on some really small AVR this can be a problem.
3. The third solution could be a mix of the 2 others: identify if an architecture provides eeprom_read/write_dword and use the optimized method in this case, otherwise use the slow/unsafe byte per byte read/write method. The implementation of eeprom_read/eeprom_write will be kept in `drivers/periph_eeprom` but with a conditional build using macro if dword fast functions are provided by an architecture. I would go for this solution but it requires a rework of your PR (and you won't have to touch AVR)

One last thing: I think the change about `_wait_for_pending_operations();` and `_IO` should be provided in another PR. This is quite orthogonal and could be merged fast.