@olegart 2018-09-28
@aabadie I can't see why we need drivers/periph_common/eeprom.c at all.