@MrKevinWeiss 2018-11-29
@olegart @aabadie is there any consensus?  I like solution 3 from @aabadie.