@aabadie 2019-09-10
This PR is partially merged in master but the speedup is still missing. I'll take over this part.